Change: Split out some of the Menu classes
